eventgo-backend/stripe-service/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
stripe>=4.0.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0,<1.5.0
psycopg2-binary>=2.9.0,<2.10.0